In-progress commit prior to major restructure

+++ fcs/control/figure-eight.c
@@ -160,7 +160,7 @@ const struct fcs_waypoint_t *end, float t) {
 
     new_point->alt = start->alt;
     new_point->airspeed = target_airspeed;
-    new_point->pitch = 4.0f * (M_PI / 180.0f);
+    new_point->pitch = 4.0f * ((float)M_PI / 180.0f);
 
     /* Always returning t means we never advance to the next path. */
     return t;
